requirements.txt
asgiref==3.11.1
coverage==7.13.4
Django==5.2.11
django-filter==25.2
django-rest-framework==0.1.0
djangorestframework==3.16.1
exceptiongroup==1.3.1
iniconfig==2.3.0
Markdown==3.10.2
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.12.0
sqlparse==0.5.5
tomli==2.4.0
typing_extensions==4.15.0
